Change the names of settings-elements
This change was made in order to make it easier to distinguish non-
settings-elements and settings-elements. One example of the problem
that was solved is the two elements "noVNC_clip" and
"noVNC_clipboard" where the first is the setting for clipping mode.
That element was now renamed to "noVNC_setting_clip".